Issue75: Allow registered users to update their ballot
Allow users to select a closing date and time for a poll in the 'Create
Poll' page.
A user can update his/her vote before the closing date of the poll.
Once a poll is closed, the user can no longer vote or update his/her
vote and sees a message: 'Sorry! This poll is closed.'